improved attached state, fix minor bugs in (de)serialization

+++ omnigibson/object_states/temperature.py
@@ -92,7 +92,7 @@ class Temperature(AbsoluteObjectState):
         self.value = state["temperature"]
 
     def _serialize(self, state):
-        return np.array([state["temperature"]])
+        return np.array([state["temperature"]], dtype=float)
 
     def _deserialize(self, state):
         return OrderedDict(temperature=state[0]), 1
